+++ archey/entries/user.py
@@ -1,8 +1,6 @@
 """User session detection class"""
 
-import os
-
-from subprocess import CalledProcessError, check_output
+import getpass
 
 from archey.entry import Entry
 
@@ -12,13 +10,9 @@ class User(Entry):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.value = os.getenv('USER') or os.getenv('LOGNAME')
-        if not self.value:
-            try:
-                self.value = check_output(
-                    ['id', '-u', '-n'],
-                    universal_newlines=True
-                ).rstrip()
-            except CalledProcessError:
-                # Should not occur, but who knows ?
-                pass
+        try:
+            self.value = getpass.getuser()
+        except ImportError:
+            # From <https://github.com/python/cpython/blob/3.9/Lib/getpass.py#L167>,
+            #  `pwd` module import _might_ fail.
+            pass